Immutable application spec (#25)
Optional immutability field tags on structs.

This is validated using reflection which iterates through all fields recursively and checks fields marked as immutable.

For slices, if they are tagged with `nais:"immutable"`, they are not allowed any change. 
If the elements of the slice are structs, the lists' elements are compared through matching "key" fields (defined on the struct's fields with the tag `nais:"key"`). 

More than one of a struct's fields can be specified as keys, and/or immutable. 
Elements of the slice with matching keys will have a full comparison.

Documentation generation is supported through `+nais:doc:Immutable=true` comments.

This PR also contains implementation of the webhook interface for the nais.io `Application` type.