Add function to apply explicit defaults to Topic.Config
We have documented some defaults for these values, but that is based
on what we experienced was the default when not specifying these in the API call creating the topic.
Aiven might change these without us noticing, and then our documented default is incorrect.
By setting them explicitly, we gain control, and can set better defaults where we think Aiven is "wrong".