Add status field for `problems`
Related to nais/system#100.

This patch adds `.status.problems` which will be used by Naiserator
to communicate error, warning and deprecation messages to the end user.

They function similarly to conditions, but are tailored for the type of
data we have inside Naiserator instead of shoehorning them into
conditions. Specifically, we can differentiate between different types
of problems without overloading the `type` field, and we are also able
to set end-of-life dates directly.

Co-authored-by: Frode Sundby <frode.sundby@nav.no>
Co-authored-by: Johnny Horvi <jhrv@users.noreply.github.com>