feat(oauth)!: separate metadata and fetching for a more idiomatic API
This changes the API for fetching metadata to a single invocation instead of
using a fluent-style API. This also moves validation of the metadata to
be allow instantiation and invocation independently of the fetching operation.

Also adds missing validation for HTTP response status code and content type
in accordance with the relevant RFCs.